parser/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
PyMuPDF==1.26.5
python-multipart==0.0.6
requests==2.31.0
pydantic==2.10.4
openai==1.58.1
python-dotenv==1.0.0
